requirements.txt
python-telegram-bot==20.7
moviepy==1.0.3
Pillow==9.5.0
python-dotenv==1.0.0
python-docx==1.0.1
python-pptx==0.6.22
PyPDF2==3.0.1
docx2pdf==0.1.8
pdf2docx==0.5.6
mammoth==1.6.0
ffmpeg-python==0.2.0
mutagen==1.47.0
piexif==1.1.3
hachoir==3.2.0
python-docx2pptx==0.2.0 